라우터 설치 및 페이지 이동
