+++ security-study/security-study-front/src/App.js
@@ -1,11 +1,19 @@
+import { BrowserRouter } from "react-router-dom";
 import Main from "./components/Main";
 
+
 function App() {
-  return (
-    <div className="App">
-        <Main />
-    </div>
-  );
-}
+
+    return (
+      <>
+        <div className="App">
+          <BrowserRouter>
+            <Main />
+          </BrowserRouter>
+        </div>
+      </>
+      
+    );
+  }
 
 export default App;
