+++ security-study/security-study-front/src/components/Join.js
@@ -1,0 +1,11 @@
+const Join = () => {
+
+    return(
+        <div>
+            회원가입 페이지
+        </div>
+    )
+
+}
+
+export default Join
